Reworked everything with Gameplay Ability System attributes, so that Health and Damage gets calculated with GAS. Refactored everything concerning Health (mostly changing int32 to float).

+++ SideScroller/Source/SideScroller/UI/HighScoreWidget.h
@@ -14,7 +14,7 @@ struct FHighScoreWidgetData
 public:
     int32 CoinAmount;
 
-    int32 HealthAmount;
+    float HealthAmount;
 
     uint32 bWonState : 1;
 };
@@ -49,7 +49,7 @@ public:
 
     // Amount of remaining health
     UPROPERTY(VisibleAnyWhere, BlueprintReadOnly, Category = "HighScoreWidget")
-    int32 HealthAmount;
+    float HealthAmount;
 
     // If the game was finished by dying or winning
     UPROPERTY(VisibleAnyWhere, BlueprintReadOnly, Category = "HighScoreWidget")
